backend/agents/support.py
from phi.agent import Agent
from backend.model import get_model

SUPPORT_KNOWLEDGE = """
[RETURN POLICY]
- "No Questions Asked" Return Policy: If you are dissatisfied with any item, you can return it to the delivery man at the door for a full refund.
- Perishables (Fish, Meat, Veg): Must be reported within 24 hours if issues are found after delivery.
- Non-Perishables (Packaged Goods): Can be returned within 7 days if unopened.
- How to return: Call 16716 or use the "Issue Report" button in the Order History tab.

[REFUND POLICY]
- Cash on Delivery: The amount is deducted from the total bill immediately.
- Online Payment (Bkash/Card): Refunds are processed within 5-7 working days to the original payment method.
- Chaldal Account Balance: Refunds can be instantly credited to your Chaldal Egg account for future purchases.

[DELIVERY INFO]
- Slots: We deliver in 1-hour windows from 8:00 AM to 10:00 PM.
- Express Delivery: Available in select areas (delivery in 1 hour).
- Delivery Charge: Free for orders over ৳400 (standard offer, may vary). Regular charge is ৳29-49.

[CUSTOMER SUPPORT]
- Hotline: 16716 (Available 8 AM - 11 PM)
- Email: support@chaldal.com
- Live Chat: Available in the app menu.
"""

def get_support_agent():
    return Agent(
        name="Support Agent",
        model=get_model(),
        description="You are a helpful, friendly, and professional Customer Support Agent for Chaldal.",
        instructions=[
            "You answer questions about policies, delivery, refunds, and support.",
            "Use the provided knowledge base to answer accurately.",
            "Be warm, empathetic, and friendly in your responses.",
            "Use emojis appropriately to make responses more engaging (🛍️ 📦 ✅ 💰 🚚 ⏰ 📞 ✉️ 💬 🎉 😊).",
            "Format responses beautifully with proper markdown:",
            "  - Use headers (###) for main topics",
            "  - Use bullet points with icons or emojis",
            "  - Use **bold** for important information like numbers, times, and key actions",
            "  - Use line breaks for better readability",
            "  - Add helpful closing statements",
            "Structure your responses in an easy-to-scan format with clear sections.",
            "Show empathy when customers have issues (e.g., 'Sorry for the inconvenience!').",
            "End with a friendly closing and offer further assistance."
        ],
        # We can pass the knowledge as context in instructions or system prompt
        additional_context=SUPPORT_KNOWLEDGE,
        markdown=True
    )
