backend/requirements.txt
fastapi
uvicorn
phidata
langchain
langchain-community
langchain-openai
supabase
psycopg2-binary
python-dotenv
openai
fastembed
